Introduce Vulkan backend implementation (`nanovg_vk`). Added Vulkan device context, texture management, pipeline creation, and descriptor handling.
